pyproject.toml
[tool.poetry]
name = "pixelalchemy"
version = "0.1.0"
description = "a website for editing images using AI (GANs)"
authors = ["rachid ouhammou <rachidouhammou21@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10,<3.12"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
pre-commit = "^3.3.3"
pyjwt = "^2.8.0"
pillow = "^10.0.0"
tensorflow-hub = "^0.14.0"
numpy = "^1.25.2"
tensorflow = "^2.13.0"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.9"
channels = "^4.0.0"
whitenoise = "^6.5.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
colorlog = "^6.7.0"


[tool.isort]
multi_line_output = 5
line_length = 119


[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
